Clean up RebalanceUtils and AdminClient
Moved a number of static methods that took an AdminClient argument
from RebalanctUtils to become members of
AdminClient.RebalancingOperations (invoked via
adminClient.rebalanceOps). Also moved createTempAdminClient to be a
static method of AdminClient.

Moved some of AdminClient.HelperOps into new inner class
AdminClient.ReplicaTypeOperations. Marked all of these to be
deprecated.

Marked a number of replicaType helper methods in RebalanceUtils as
deprecated and annotated with a TODO as well.